// src/google-auth/dtos/google-user.dto.ts
export class GoogleUserDto {
  profile: {
    id: string;
    displayName: string;
    name: {
      familyName: string;
      givenName: string;
    };
    emails: [
      {
        value: string;
        verified: boolean;
      },
    ];
    photos: [
      {
        value: string;
      },
    ];
    provider: string;
    _raw: string;
    _json: {
      sub: string;
      name: string;
      given_name: string;
      family_name: string;
      picture: string;
      email: string;
      email_verified: boolean;
      locale: string;
    };
  };
  accessToken: string;
}